namespaces: Don't fully resolve types
The `--namespaces` option _appeared_ to be showing duplicate values. However,
what was actually happening is that the code which resolves the symlinks
the entries in `/proc/self/ns/` point to was only showing the name of the
symlink target, not the name of the link. This wasn't a problem previously
since the values were identical. However, with the advent of the
`pid_for_children` and `time_for_children` namespaces (Linux 5.6), it is a
problem.

Here's an example of what `--namespaces` used to show:

```
namespaces:
  cgroup: 4026531835
  ipc: 4026531839
  mnt: 4026531840
  net: 4026532008
  pid: 4026531836
  pid: 4026531836
  time: 4026531834
  time: 4026531834
  user: 4026531837
  uts: 4026531838
```

(Notice the double entries for `pid` and `time`!)

Whereas here's an example of what is now displayed:

```
namespaces:
  cgroup: cgroup:[4026531835]
  ipc: ipc:[4026531839]
  mnt: mnt:[4026531840]
  net: net:[4026532008]
  pid: pid:[4026531836]
  pid_for_children: pid:[4026531836]
  time: time:[4026531834]
  time_for_children: time:[4026531834]
  user: user:[4026531837]
  uts: uts:[4026531838]
```

Note that the format has changed slightly (it seemed pointless to remove the
square brackets).

See `namespaces(7)` for further details.

Signed-off-by: James O. D. Hunt <jamesodhunt@gmail.com>