CI: Disable homebrew for OSX
Homebrew is borked in Travis' OSX environment; updating `brew` triggers
openjdk to be built from source, which times out after 10 minutes of no
output.

Since we don't want Java and the Travis config never requested Java,
this is a pita. Therefore force homebrew to be disabled. The downsides of
this change are that on OSX:

- The `pr_list` unit tests are won't run (since the `libcheck` library
  won't be available).
- The XML and JSON tests in `src/tests/check_all_args` won't be run
  since `xmlwf(1)` and `json_pp(1)` (which is part of the `perl(1)`
  package) won't be available.

Signed-off-by: James O. D. Hunt <jamesodhunt@gmail.com>